0.0.3.1

+++ functions/delete-mod.php
@@ -6,10 +6,14 @@ if(empty($_GET['id'])){
 	die("Id not specified.");
 }
 if(!$_SESSION['user']||$_SESSION['user']=="") {
-	die("You need to be logged in!");
+	die("Unauthorized require or login session has expired!");
 }
-$modq = mysqli_query($conn, "SELECT * FROM `mods` WHERE `id` = '".$_GET['id']."'");
+if(substr($_SESSION['perms'],4,1)!=="1") {
+	echo 'Insufficient permission!';
+	exit();
+}
+$modq = mysqli_query($conn, "SELECT * FROM `mods` WHERE `id` = '".mysqli_real_escape_string($conn,$_GET['id'])."'");
 $mod = mysqli_fetch_array($modq);
 unlink("../".$mod['type']."s/".$mod['filename']);
-mysqli_query($conn, "DELETE FROM `mods` WHERE `id` = '".$_GET['id']."'");
+mysqli_query($conn, "DELETE FROM `mods` WHERE `id` = '".mysqli_real_escape_string($conn,$_GET['id'])."'");
 exit();
